rename checklevel package gh_control
