+++ sourcetool/cmd/checklevel.go
@@ -10,7 +10,7 @@ import (
 	"os"
 
 	"github.com/slsa-framework/slsa-source-poc/sourcetool/pkg/attest"
-	"github.com/slsa-framework/slsa-source-poc/sourcetool/pkg/checklevel"
+	"github.com/slsa-framework/slsa-source-poc/sourcetool/pkg/gh_control"
 
 	"github.com/google/go-github/v68/github"
 	"github.com/spf13/cobra"
@@ -44,7 +44,7 @@ func doCheckLevel(commit, owner, repo, branch, outputVsa, outputUnsignedVsa stri
 	gh_client := github.NewClient(nil)
 	ctx := context.Background()
 
-	sourceLevel, err := checklevel.DetermineSourceLevelControlOnly(ctx, gh_client, commit, owner, repo, branch)
+	sourceLevel, err := gh_control.DetermineSourceLevelControlOnly(ctx, gh_client, commit, owner, repo, branch)
 	if err != nil {
 		log.Fatal(err)
 	}
